reorganized and added testing
