+++ src/util/operationValidationUtils.js
